Fix bug with 1/8 representation
We were using the remainder of the decimal part in a dictionary to
detect whether we had already seen the same state in a previous
iteration of the algorithm.

However, in the case of 1/8, the algorithm could do:

1/8 -> 10/8=5/4 -> 1/4, with the 1 remainder repeating, but 1/4 != 1/8.

We use the full division instead as the key in the dictionary.